bugs bugs and more bugs

+++ src/PCL/OKHOSTING.UI/Controls/Forms/IntegerField.cs
@@ -22,7 +22,7 @@ namespace OKHOSTING.UI.Controls.Forms
 			}
 			set
 			{
-				ValueControl.Value = ((int) value).ToString();
+				ValueControl.Value = value.ToString();
 			}
 		}
